+++ frontend/src/app/services/ordinals/cat21.service.helper.jest.ts
@@ -1,5 +1,6 @@
 /* eslint-disable no-console */
 import { describe, expect, it } from '@jest/globals';
+
 import { createRandomPrivateKey, getMinimumUtxoSize } from './cat21.service.helper';
 
 describe('getMinimumUtxoSize', () => {
